test: extract cpu template utils into a file
